+++ coaster/sqlalchemy/__init__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+
+"""
+SQLAlchemy patterns
+===================
+
+Coaster provides a number of SQLAlchemy helper functions and mixin classes
+that add standard columns or special functionality.
+
+All functions and mixins are importable from the :mod:`coaster.sqlalchemy`
+namespace.
+"""
+
+from __future__ import absolute_import
+
+from .comparators import *  # NOQA
+from .functions import *  # NOQA
+from .roles import *  # NOQA
+from .annotations import *  # NOQA
+from .immutable_annotation import *  # NOQA
+from .mixins import *  # NOQA
+from .columns import *  # NOQA